hw for lesson 1
